Add a script to build a test case coverage report.
By default it produces a report file with a source file level summary of
coverage, with percentage of lines executed.

The `-s` option will instead print a summary to standard out detailing
all the functions executed with execution counts per line.

A regex argument to `-s` may be supplied to filter which functions are
in the summary.