Remove the `yurtc` binary from the coverage report.
It's no longer needed to get a full coverage summary as we're linking
against the `yurtc` library in `tests.rs` rather than calling `yurtc`
externally.

It seems the output is still a little bit confused as the unit tests and
integration tests are still reported separately, I think.  But
everything is counted.