Use the /tmp dir for the working files.
Something strange was going on when working relative to the current
directory.  If the script had to rebuild the binaries via `cargo test`
it would leave the working dir behind.  Cargo seemed to be changing the
current directory or something.

Making the working dir absolute seems to have fixed it.  And /tmp is as
good a place as any to put it.