Reorganise project parsing a bit.
The macro decls and calls are now stored in the `ProjectParser` so they
can be expanded after source parsing is done.

The `ProjectParser` methods are all less recursive now, using a 'pending
path' stack, which means modules are parsed in reverse order.  If this a
problem (shouldn't be) then we could make it a pending path queue.