Add macro call parsing.
Unfortunately at this stage it's broken, as the macro decls and calls
are stored outside of the II since they require the 'sc lifetime due to
hanging on to tokens and the II does not.

But this means they're thrown away at the end of parsing each module,
before we've expanded them.  We can't expand before they're thrown away
since they may be referred to from a separate module.