Add a flag to the LALRPOP build script to make it less aggressive.
Until now it would try to rebuild the parser when _any_ file was
updated, including any source or even test file.  It usually meant
nothing changed if the grammar file was unchanged, but it still took a
few seconds, which is especially frustrating when iterating on tests.

I haven't tested this longer term -- I hope it doesn't make it harder to
update the grammar.