Rework generator unrolling to be less brittle.
Use the root set to find generators rather than all expressions and then
order them explicitly to be sure nested inner generators are unrolled
before the outer.

Also just replace the generator expression with the unrolled expression
rather than deleting it.  The removal was a bit blunt and didn't check
for potential other 'uses'.

And:
- Add `IntermediateIntent::visitor()`.
- Remove `IntermediateIntent::remove_expr()`.
- Use `IntermediateIntent::Expr` iterator in `scalarize.rs`.
- Improve error recovery for array comparison lowering in
  `scalarize.rs`.