feat: Add support for dependencies to `pintc` in the parser
The `ProjectParser` now stores `deps` internally in the form of a map
from their name to their entry point and parsed program.

The `deps` are used during `find_next_path` to determine whether or not
the path may refer to some dependency.

During path analysisi, if a dependency path is encountered, we add its
path to `pending_paths`.

Dependencies are treated as though they are modules that are always
present within the global namespace.

Also updates some test and bin usages of `parse_project` to account for
providing dependencies.

In the case that a local module name would conflict with a dependency
name, currently we produce the `DualModulity` error. This is a temporary
solution pending the addition of a `root::` namespace, which hasn't been
included as a part of this PR in order to keep the changes minimal.