feat: pint-pkg supports contract deps, includes a simple test
This updates the `pint-pkg` build process to build a single small
library for each built contract.

This library is a small temporary directory with a `lib.pnt` entry point
containing the `ADDRESS` for the contract, along with a submodule and
`ADDRESS` for each intent.

E.g. the counter app generates:

```
<temp_dir>/<contract-content-address>/lib.pnt
<temp_dir>/<contract-content-address>/Init.pnt
<temp_dir>/<contract-content-address>/Increment.pnt
```

where

- `lib.pnt` contains the contract content address
- `Init.pnt` is the `::Init` intent's content address
- `Increment.pnt` is the `::Increment` intent's content address

Currently we use `var` for these consts, but these should be updated
when addressing #584.

A test is included where a contract `foo` depends on a contract `bar`
and uses `bar`'s addresses within some simple constraints. The test
checks that `bar`'s CAs all appear within `foo`'s constraint bytecode.