refactor: Change `pint-abi-gen` to use `pint-abi-visit` for key construct
This changes the `pint-abi-gen` crate to ignore the deserialized ABI
keys in favour of constructing them based on the `pint-abi-visit`
traversal.

This is achieved by tracking how each node in the `KeyedVarTree` is
"nested" within the set of keyed vars.

This required some changes to `pint-abi-visit` to allow for indexing
into "child" nodes for method generation.