Tweak the `[[` and `]]` parsing.
A fix was needed for parsing array expressions starting with `[[`.  To
simplify the fix the `]]` token was removed and the grammar will just
parse `]` `]` for instances like `ctrct[[c]]::pred[[p]](..)`.

The only issue introduce here is array types which array index exprs for
ranges.

E.g., `let blah: int[[1,2,3][1]];` which optimises to `let blah: int[2];`

The parser sees `[[` and barfs.  This is a very rare and weird edge case
though, so the tests just use `let blah: int[ [1,2,3][1]];` instead.