Add ASM gen for a map over a fixed size array expression.
The array expression is on the stack.  The map will then move it to the
heap before mapping to a different buffer in the heap.  This is quite
inefficient and to be improved requires slight changes to the VM, OR for
array expressions to never be on the stack in the first place.